Enable warnings in nvcc compilation
Fix float rintf call.

+++ hoomd/BoxDim.h
@@ -299,7 +299,7 @@ BoxDim
             #ifdef __HIPCC__
             if (m_periodic.z)
                 {
-                Scalar img = rintf(w.z * m_Linv.z);
+                Scalar img = rint(w.z * m_Linv.z);
                 w.z -= L.z * img;
                 w.y -= L.z * m_yz * img;
                 w.x -= L.z * m_xz * img;
@@ -307,17 +307,17 @@ BoxDim
 
             if (m_periodic.y)
                 {
-                Scalar img = rintf(w.y * m_Linv.y);
+                Scalar img = rint(w.y * m_Linv.y);
                 w.y -= L.y * img;
                 w.x -= L.y * m_xy * img;
                 }
 
             if (m_periodic.x)
                 {
-                w.x -= L.x * rintf(w.x * m_Linv.x);
+                w.x -= L.x * rint(w.x * m_Linv.x);
                 }
             #else
-            // on the cpu, branches are faster than calling rintf
+            // on the cpu, branches are faster than calling rint
             if (m_periodic.z)
                 {
                 if (w.z >= m_hi.z)
